aidersplatform/requirements.txt
affine==2.3.0
altgraph==0.17
aniso8601==8.1.1
argh==0.26.2
asgiref==3.5.0
atomicwrites==1.4.0
attrs==20.3.0
autobahn==20.12.3
Automat==20.2.0
backports.statistics==0.1.0
backports.zoneinfo==0.2.1
bcrypt==3.2.0
beautifulsoup4==4.10.0
catkin-pkg==0.4.23
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
click==7.1.2
click-plugins==1.1.1
cligj==0.7.1
colorama==0.4.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.3.1
cycler==0.10.0
Cython==0.29.21
dataclasses==0.6
ddtrace==0.58.0
decorator==4.4.2
disp==0.0.3
distro==1.5.0
Django==4.0.3
django-crispy-forms==1.14.0
django-floppyforms==1.9.0
django-formtools==2.3
django-guardian==2.4.0
django-livereload-server==0.4
django-map-widgets==0.4.0
djangorestframework==3.13.1
django-extensions
docker==5.0.0
docopt==0.6.2
docutils==0.16
drf-chunked-upload==0.5.1
drf-yasg==1.20.0
earthpy==0.9.2
ExifRead==2.3.2
factory-boy==3.2.1
Faker==13.0.0
filelock==3.4.0
Fiona==1.8.18
flock==0.1
fonttools==4.29.1
future==0.18.2
geographiclib==1.50
geojson==2.5.0
geopandas==0.8.2
geopy==1.23.0
hyperlink==21.0.0
idna==2.10
image==1.5.33
imageio==2.9.0
importlib-metadata==3.10.0
incremental==17.5.0
inflection==0.5.1
inotify==0.2.10
ipython-genutils==0.2.0
iso8601==1.0.0
itypes==1.2.0
jeepney==0.6.0
Jinja2==2.11.3
joblib==1.0.0
json2html==1.3.0
jsonschema==3.2.0
keyring==23.0.1
kiwisolver==1.3.1
livehttp==0.0.7
MarkupSafe==1.1.1
matplotlib==3.5.1
more-itertools==8.12.0
mpmath==1.2.1
munch==2.5.0
mysql-connector-python==8.0.28
mysqlclient==2.1.0
nbformat==5.5.0
netifaces==0.10.9
networkx==2.5
nose==1.3.7
numpy==1.19.5
openapi-codec==1.3.2
opencv-python==4.5.1.48
packaging==21.3
pandas==1.1.5
pathtools==0.1.2
pexpect==4.8.0
piexif==1.1.3
pigar==0.10.0
Pillow==9.0.0
pipreqs==0.4.11
pluggy==0.13.1
pluginlib==0.8.1
priority==1.3.0
protobuf==3.19.4
psutil==5.8.0
ptyprocess==0.7.0
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycrypto==2.6.1
PyExifTool==0.1.1
PyHamcrest==2.0.2
pymongo==3.11.3
PyMySQL==1.0.2
pyodm==1.5.9
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyproj==3.0.0.post1
pymavlink==2.4.29
PyQt5==5.15.5
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pyqtgraph==0.12.3
pyrsistent==0.17.3
pyserial==3.5
pytest==5.2.1
python-dateutil==2.8.1
python-dotenv==0.19.2
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
requests==2.24.0
requests-toolbelt==0.9.1
rioxarray==0.3.1
rospkg==1.2.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
schedule==1.0.0
scikit-image==0.17.2
scikit-learn==0.24.1
scipy==1.5.4
SecretStorage==3.3.1
service-identity==18.1.0
setuptools-svn==1.2
Shapely==1.7.1
simplejson==3.17.6
sklearn==0.0
snuggs==1.4.7
soupsieve==2.3.1
sqlparse==0.4.2
sympy==1.9
telemetry==0.3.6
tenacity==8.0.1
threadpoolctl==2.1.0
tinyaes==1.0.1
tornado==6.1
tqdm==4.56.0
traitlets==5.1
transitions==0.8.6
Twisted==20.3.0
txaio==20.12.1
typing-extensions==3.7.4.3
ua-parser==0.10.0
uritemplate==4.1.1
urllib3==1.25.11
user-agents==2.2.0
utility==1.0
vincenty==0.1.4
watch==0.2.7
wcwidth==0.2.5
websocket-client==0.58.0
Werkzeug==1.0.1
wincertstore==0.2
xarray==0.16.2
yarg==0.1.9
zipp==3.4.1
zope.interface==5.2.0
numba==0.48
cython-bbox==0.1.3
tdqm==0.0.1
seaborn==0.12.1
onnx==1.12.0
imutils==0.5.4
torch==1.11.0
torchvision==0.12.0
cupy-cuda113==10.6.0
drf-extra-fields==3.4.0
psutil==5.8.0
GPUtil==1.4.0 
channels==3.0.4
open3d==0.16.0
itsdangerous==2.0.1
reportlab==3.6.12
pdfkit==1.0.0